feat/MSSDK-1545: move handleUnsubscribe() to Unsubscribe.tsx
